Move Simulation.Counters utility class to SpringSaLaDExporter
